Vaccinations for COVID-19 Counts for US Counties from Center for Disease Control (CDC)

This functions pulls vaccine data for US counties from the CDC. No data is stored within the package, so the data should continuously update over time as long as the links do not change.

cdc_vac_us_by_cnty()

Value

A dataframe/tibble

Details

Website:

- CDC COVID Tracker: https://covid.cdc.gov/covid-data-tracker/#datatracker-home

- CDC State Site: https://data.cdc.gov/Vaccinations/COVID-19-Vaccinations-in-the-United-States-County/8xkx-amqh

- Data: https://data.cdc.gov/api/views/8xkx-amqh/rows.csv?accessType=DOWNLOAD

Examples

vac_by_cnty <- R.COVID.19::cdc_vac_us_by_cnty()

#> Rows: 761182 Columns: 27

#> -- Column specification -------------------------------------------------------------------------
#> Delimiter: ","
#> chr  (9): Date, FIPS, Recip_County, Recip_State, SVI_CTGY, Series_Complete_P...
#> dbl (18): MMWR_week, Series_Complete_Pop_Pct, Series_Complete_Yes, Series_Co...

#> 
#> 
#> i Use `spec()` to retrieve the full column specification for this data.
#> i Specify the column types or set `show_col_types = FALSE` to quiet this message.

head(vac_by_cnty)

#> # A tibble: 6 x 27
#>   Date       FIPS  MMWR_week Recip_County Recip_State Series_Complete~
#>   <date>     <chr>     <dbl> <chr>        <chr>                  <dbl>
#> 1 2020-12-13 02013        51 Aleutians E~ AK                         0
#> 2 2020-12-14 02013        51 Aleutians E~ AK                         0
#> 3 2020-12-15 02013        51 Aleutians E~ AK                         0
#> 4 2020-12-16 02013        51 Aleutians E~ AK                         0
#> 5 2020-12-17 02013        51 Aleutians E~ AK                         0
#> 6 2020-12-18 02013        51 Aleutians E~ AK                         0
#> # ... with 21 more variables: Series_Complete_Yes <dbl>,
#> #   Series_Complete_12Plus <dbl>, Series_Complete_12PlusPop_Pct <dbl>,
#> #   Series_Complete_18Plus <dbl>, Series_Complete_18PlusPop_Pct <dbl>,
#> #   Series_Complete_65Plus <dbl>, Series_Complete_65PlusPop_Pct <dbl>,
#> #   Completeness_pct <dbl>, Administered_Dose1_Recip <dbl>,
#> #   Administered_Dose1_Pop_Pct <dbl>, Administered_Dose1_Recip_12Plus <dbl>,
#> #   Administered_Dose1_Recip_12PlusPop_Pct <dbl>,
#> #   Administered_Dose1_Recip_18Plus <dbl>,
#> #   Administered_Dose1_Recip_18PlusPop_Pct <dbl>,
#> #   Administered_Dose1_Recip_65Plus <dbl>,
#> #   Administered_Dose1_Recip_65PlusPop_Pct <dbl>, SVI_CTGY <chr>,
#> #   Series_Complete_Pop_Pct_SVI <chr>, Series_Complete_12PlusPop_Pct_SVI <chr>,
#> #   Series_Complete_18PlusPop_Pct_SVI <chr>,
#> #   Series_Complete_65PlusPop_Pct_SVI <chr>