COVID-19 US Testing Counts from CovidTracking.com

This functions pulls US testing counts of COVID-19 data from covidtracking.com. The data is tidyed so that it is more user friendly. No data is stored within the package, so the data should continuously update over time as long as the links do not change.

covtrck_states_testing()

Value

A tsibble/dataframe/tibble

Details

Website: https://covidtracking.com/api

Deprecated fields The following fields are deprecated and are no longer being updated with new data.

/api/v1/states/current.json:

positiveScore negativeScore negativeRegularScore commercialScore score grade - Use dataQualityGrade instead total

/api/v1/us/current.json

posNeg total

Examples

tests_daily <- R.COVID.19::covtrck_states_testing()

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   state = col_character(),
#>   dataQualityGrade = col_character(),
#>   lastUpdateEt = col_character(),
#>   hash = col_character(),
#>   dateChecked = col_datetime(format = ""),
#>   fips = col_character()
#> )

#> See spec(...) for full column specifications.

#> `mutate_if()` ignored the following grouping variables:
#> Column `fips`

#> `mutate_if()` ignored the following grouping variables:
#> Column `fips`

head(tests_daily)

#> # A tibble: 6 x 30
#> # Groups:   fips [1]
#>   greg_d       date fips  state total_tests positive negative pending
#>   <date>      <dbl> <chr> <chr>       <dbl>    <dbl>    <dbl>   <dbl>
#> 1 2020-03-07 2.02e7 01    AL              0        0        0       0
#> 2 2020-03-08 2.02e7 01    AL              0        0        0       0
#> 3 2020-03-09 2.02e7 01    AL              0        0        0       0
#> 4 2020-03-10 2.02e7 01    AL              0        0        0       0
#> 5 2020-03-11 2.02e7 01    AL             10        0       10       0
#> 6 2020-03-12 2.02e7 01    AL             10        0       10       0
#> # ... with 22 more variables: total_tests_increase <dbl>,
#> #   pendingIncrease <dbl>, positiveIncrease <dbl>, negativeIncrease <dbl>,
#> #   totalTestResults <dbl>, totalTestResultsIncrease <dbl>, recovered <dbl>,
#> #   recoveredIncrease <dbl>, death <dbl>, deathIncrease <dbl>,
#> #   hospitalized <dbl>, hospitalizedIncrease <dbl>,
#> #   hospitalizedCurrently <dbl>, hospitalizedCumulative <dbl>,
#> #   inIcuCurrently <dbl>, inIcuCumulative <dbl>, onVentilatorCurrently <dbl>,
#> #   onVentilatorCumulative <dbl>, dataQualityGrade <chr>, lastUpdateEt <chr>,
#> #   hash <chr>, dateChecked <dttm>