FINRA Shelter in Place Dates

This functions pulls dates for shelter in place order by states with notes and start date and end dates, as well as, extention dates from the FINRA COVID-19 website. No Parameters

finra_shelter_in_place()

Value

A dataframe/tibble

Details

Website: https://www.finra.org/rules-guidance/key-topics/covid-19/shelter-in-place

Examples

stay_orders <- R.COVID.19::finra_shelter_in_place()

#> Warning:  7 failed to parse.

head(stay_orders)

#> # A tibble: 6 x 7
#>   state  order   order_beg_d order_end_org order_end_d order_extended re_opening
#>   <chr>  <chr>   <date>      <chr>         <date>               <dbl>      <dbl>
#> 1 Alaba~ ""      2020-04-03  4/30/2020     2020-04-30               0          1
#> 2 Alaska Alaska~ 2020-03-27  Rescinded Ap~ 2020-04-24               0          1
#> 3 Arizo~ Execut~ 2020-03-30  Extended to ~ 2020-05-15               1          1
#> 4 Calif~ ""      2020-03-19  None          NA                       0          1
#> 5 Color~ Amende~ 2020-03-25  Extended to ~ 2020-04-26               1          1
#> 6 Conne~ Guidan~ 2020-03-20  Extended to ~ 2020-05-20               1          1