Vaccinations for COVID-19 Counts for US States from Our World in Data

This functions pulls COVID-19 vaccine data for US states from Our World in Data website GitHub repository and downloads it. No data is stored within the package, so the data should continuously update over time as long as the links do not change.

owid_vac_us_by_state()

Value

A dataframe/tibble

Details

Website:

- https://github.com/owid/covid-19-data/tree/master/public/data/vaccinations

- https://raw.githubusercontent.com/owid/covid-19-data/master/public/data/vaccinations/us_state_vaccinations.csv

Examples

confirmed_st_daily <- R.COVID.19::owid_vac_us_by_state()

#> Parsed with column specification:
#> cols(
#>   date = col_date(format = ""),
#>   location = col_character(),
#>   total_vaccinations = col_double(),
#>   total_distributed = col_double(),
#>   people_vaccinated = col_double(),
#>   people_fully_vaccinated_per_hundred = col_double(),
#>   total_vaccinations_per_hundred = col_double(),
#>   people_fully_vaccinated = col_double(),
#>   people_vaccinated_per_hundred = col_double(),
#>   distributed_per_hundred = col_double(),
#>   daily_vaccinations_raw = col_double(),
#>   daily_vaccinations = col_double(),
#>   daily_vaccinations_per_million = col_double(),
#>   share_doses_used = col_double()
#> )

head(confirmed_st_daily)

#> # A tibble: 6 x 14
#>   date       location total_vaccinati~ total_distribut~ people_vaccinat~
#>   <date>     <chr>               <dbl>            <dbl>            <dbl>
#> 1 2021-01-12 Alabama             78134           377025            70861
#> 2 2021-01-13 Alabama             84040           378975            74792
#> 3 2021-01-14 Alabama             92300           435350            80480
#> 4 2021-01-15 Alabama            100567           444650            86956
#> 5 2021-01-16 Alabama                NA               NA               NA
#> 6 2021-01-17 Alabama                NA               NA               NA
#> # ... with 9 more variables: people_fully_vaccinated_per_hundred <dbl>,
#> #   total_vaccinations_per_hundred <dbl>, people_fully_vaccinated <dbl>,
#> #   people_vaccinated_per_hundred <dbl>, distributed_per_hundred <dbl>,
#> #   daily_vaccinations_raw <dbl>, daily_vaccinations <dbl>,
#> #   daily_vaccinations_per_million <dbl>, share_doses_used <dbl>