COVID-19 Public Health Status by County shapefiles from ESRI

This functions pulls the public health status by county in the US as of the current date with the shapefiles, which are saved to the current working directory. The data comes from ESRI open data website. The data is 1 row per county.

pub_hlth_status_by_cnty_shp()

Value

A dataframe/tibble. This functions has side effects in that it downloads shapefiles corresponding to the counties to the current working directory.

Details

Website: https://coronavirus-disasterresponse.hub.arcgis.com/datasets/97792521be744711a291d10ecef33a61/data?geometry=76.921

Examples

pub_status_cnty_shp <- R.COVID.19::pub_hlth_status_by_cnty_shp()

#> Reading layer `COVID19_Public_Health_Emergency_Status_by_County' from data source `C:\Users\marqu\OneDrive\Documents\GitHub\R.COVID.19\docs\reference\COVID19_Public_Health_Emergency_Status_by_County.shp' using driver `ESRI Shapefile'
#> Simple feature collection with 3152 features and 28 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -178.2176 ymin: -14.37427 xmax: 146.0827 ymax: 71.40624
#> CRS:            4326

head(pub_status_cnty_shp)

#> Simple feature collection with 6 features and 28 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -80.83996 ymin: 34.55392 xmax: -76.02569 ymax: 36.55118
#> CRS:            4326
#>   objectid_1 objectid     name     state_name  fips   sqmi cnty_name county_st
#> 1       3001     2659  Sampson North Carolina 37163 947.21   Sampson        NC
#> 2       3002     2660 Scotland North Carolina 37165 320.25  Scotland        NC
#> 3       3003     2661   Stanly North Carolina 37167 404.28    Stanly        NC
#> 4       3004     2662   Stokes North Carolina 37169 455.72    Stokes        NC
#> 5       3005     2663  Tyrrell North Carolina 37177 390.17   Tyrrell        NC
#> 6       3006     2664    Union North Carolina 37179 639.47     Union        NC
#>       county_st_ ecode                                                 chapter
#> 1 North Carolina 33132              The ARC of the Sandhills of North Carolina
#> 2 North Carolina 33132              The ARC of the Sandhills of North Carolina
#> 3 North Carolina 33068 The ARC serving the Southern Piedmont of North Carolina
#> 4 North Carolina 33188    The ARC serving the Piedmont Triad of North Carolina
#> 5 North Carolina 33380                  The ARC of Northeastern North Carolina
#> 6 North Carolina 33068 The ARC serving the Southern Piedmont of North Carolina
#>   rcode                        region dcode              division time_zone
#> 1 33R16 Eastern North Carolina Region   D26 Mid-Atlantic Division   Eastern
#> 2 33R16 Eastern North Carolina Region   D26 Mid-Atlantic Division   Eastern
#> 3 33R08      Greater Carolinas Region   D26 Mid-Atlantic Division   Eastern
#> 4 33R08      Greater Carolinas Region   D26 Mid-Atlantic Division   Eastern
#> 5 33R16 Eastern North Carolina Region   D26 Mid-Atlantic Division   Eastern
#> 6 33R08      Greater Carolinas Region   D26 Mid-Atlantic Division   Eastern
#>   fema_regio                                             notes
#> 1         IV Governor issued statewide shelter in place orders
#> 2         IV Governor issued statewide shelter in place orders
#> 3         IV Governor issued statewide shelter in place orders
#> 4         IV Governor issued statewide shelter in place orders
#> 5         IV Governor issued statewide shelter in place orders
#> 6         IV Governor issued statewide shelter in place orders
#>                 last_updat                        local_heal  dro
#> 1 2020-03-28T00:57:05.000Z Govt Ordered Community Quarantine <NA>
#> 2 2020-03-28T00:57:05.000Z Govt Ordered Community Quarantine <NA>
#> 3 2020-03-28T00:57:05.000Z Govt Ordered Community Quarantine <NA>
#> 4 2020-03-28T00:57:05.000Z Govt Ordered Community Quarantine <NA>
#> 5 2020-03-28T00:57:05.000Z Govt Ordered Community Quarantine <NA>
#> 6 2020-03-28T00:57:05.000Z Govt Ordered Community Quarantine <NA>
#>                 created_da    created_us               last_edite    last_edi_1
#> 1 2020-04-02T16:09:30.670Z alow_rconline 2020-04-02T16:09:30.670Z alow_rconline
#> 2 2020-04-02T16:09:30.670Z alow_rconline 2020-04-02T16:09:30.670Z alow_rconline
#> 3 2020-04-02T16:09:30.670Z alow_rconline 2020-04-02T16:09:30.670Z alow_rconline
#> 4 2020-04-02T16:09:30.670Z alow_rconline 2020-04-02T16:09:30.670Z alow_rconline
#> 5 2020-04-02T16:09:30.670Z alow_rconline 2020-04-02T16:09:30.670Z alow_rconline
#> 6 2020-04-02T16:09:30.670Z alow_rconline 2020-04-02T16:09:30.670Z alow_rconline
#>                               globalid Shape__Are Shape__Len
#> 1 3dcfbfc7-0e3d-401b-88f7-5f7249ea89de 3640693856   279006.3
#> 2 a5b6e8f9-5901-4ea6-a409-bfac98b25c38 1203439364   147599.9
#> 3 f1893887-95e9-4bbc-a28d-d82826094823 1570936706   181414.4
#> 4 957a2f73-df85-4bb6-82b2-14182e4cd878 1906940465   174859.0
#> 5 d7ff6071-5990-4f92-8840-e79fc7eefd8b 1506027818   173913.2
#> 6 a3a88ae3-fb02-4f1f-b708-b6e5bba3a52f 2461669771   208739.1
#>                         geometry
#> 1 MULTIPOLYGON (((-78.11348 3...
#> 2 MULTIPOLYGON (((-79.45574 3...
#> 3 MULTIPOLYGON (((-80.07119 3...
#> 4 MULTIPOLYGON (((-80.02544 3...
#> 5 MULTIPOLYGON (((-76.16695 3...
#> 6 MULTIPOLYGON (((-80.83996 3...