COVID-19 US Death Counts from John Hopkins

This functions pulls US deaths COVID-19 data from John Hopkins GitHub repository downloads it and transpose it so that it is tidy. No data is stored within the package, so the data should continuously update over time as long as the links do not change.

us_geo_deaths_daily()

Value

A dataframe/tibble

Details

Website: https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_deaths_US.csv

Examples

deaths_daily <- R.COVID.19::us_geo_deaths_daily()

#> Parsed with column specification:
#> cols(
#>   .default = col_double(),
#>   iso2 = col_character(),
#>   iso3 = col_character(),
#>   Admin2 = col_character(),
#>   Province_State = col_character(),
#>   Country_Region = col_character(),
#>   Combined_Key = col_character()
#> )

#> See spec(...) for full column specifications.

head(deaths_daily)

#> # A tibble: 6 x 14
#>      UID iso2  iso3  code3  FIPS Admin2 Province_State Country_Region   Lat
#>    <dbl> <chr> <chr> <dbl> <dbl> <chr>  <chr>          <chr>          <dbl>
#> 1 8.40e7 US    USA     840  1001 Autau~ Alabama        US              32.5
#> 2 8.40e7 US    USA     840  1001 Autau~ Alabama        US              32.5
#> 3 8.40e7 US    USA     840  1001 Autau~ Alabama        US              32.5
#> 4 8.40e7 US    USA     840  1001 Autau~ Alabama        US              32.5
#> 5 8.40e7 US    USA     840  1001 Autau~ Alabama        US              32.5
#> 6 8.40e7 US    USA     840  1001 Autau~ Alabama        US              32.5
#> # ... with 5 more variables: Long_ <dbl>, Combined_Key <chr>, Population <dbl>,
#> #   greg_d <chr>, deaths_cases <dbl>