COVID-19 Public Health Status by County shapefiles from ESRI

This functions pulls the public health status by county in the US as of the current date with the shapefiles, which are saved to the current working directory. The data comes from ESRI open data website. The data is 1 row per county.

pub_hlth_status_by_cnty_shp()

Value

A dataframe/tibble. This functions has side effects in that it downloads shapefiles corresponding to the counties to the current working directory.

Details

Website: https://coronavirus-disasterresponse.hub.arcgis.com/datasets/97792521be744711a291d10ecef33a61/data?geometry=76.921

Examples

pub_status_cnty_shp <- R.COVID.19::pub_hlth_status_by_cnty_shp()

#> Reading layer `8acac48c-57be-4cc9-92a2-b932b279b46c2020329-1-gva0h9.tiu36' from data source `C:\Users\marqu\OneDrive\Documents\GitHub\R.COVID.19\docs\reference\8acac48c-57be-4cc9-92a2-b932b279b46c2020329-1-gva0h9.tiu36.shp' using driver `ESRI Shapefile'
#> Simple feature collection with 3152 features and 28 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -178.2176 ymin: -14.37427 xmax: 146.0827 ymax: 71.40624
#> CRS:            4326

head(pub_status_cnty_shp)

#> Simple feature collection with 6 features and 28 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -178.2176 ymin: 47.24046 xmax: -96.83707 ymax: 61.49217
#> CRS:            4326
#>   objectid_1 objectid           name   state_name  fips    sqmi
#> 1          1        1 Aleutians East       Alaska 02013 7036.89
#> 2          2     1201         Traill North Dakota 38097  862.51
#> 3          3     1202          Walsh North Dakota 38099 1294.14
#> 4          4        2 Aleutians West       Alaska 02016 3597.77
#> 5          5     1203           Ward North Dakota 38101 2055.85
#> 6          6        3      Anchorage       Alaska 02020 1733.41
#>                    cnty_name county_st   county_st_ ecode
#> 1     Aleutians East Borough        AK       Alaska 02012
#> 2                     Traill        ND North Dakota 34118
#> 3                      Walsh        ND North Dakota 34118
#> 4 Aleutians West Census Area        AK       Alaska 02012
#> 5                       Ward        ND North Dakota 34204
#> 6     Anchorage Municipality        AK       Alaska 02012
#>                                                    chapter rcode         region
#> 1                                       ARC serving Alaska 02R04  Alaska Region
#> 2 ARC serving Eastern North Dakota and Northwest Minnesota 34R08 Dakotas Region
#> 3 ARC serving Eastern North Dakota and Northwest Minnesota 34R08 Dakotas Region
#> 4                                       ARC serving Alaska 02R04  Alaska Region
#> 5                         ARC serving Western North Dakota 34R08 Dakotas Region
#> 6                                       ARC serving Alaska 02R04  Alaska Region
#>   dcode               division time_zone fema_regio notes last_updat
#> 1   D21       Pacific Division    Alaska          X  <NA> 2020-03-28
#> 2   D23 North Central Division   Central       VIII  <NA> 2020-03-17
#> 3   D23 North Central Division   Central       VIII  <NA> 2020-03-17
#> 4   D21       Pacific Division    Alaska          X  <NA> 2020-03-28
#> 5   D23 North Central Division  Mountain       VIII  <NA> 2020-03-17
#> 6   D21       Pacific Division    Alaska          X  <NA> 2020-03-28
#>                          local_heal  dro created_da    created_us last_edite
#> 1 Govt Ordered Community Quarantine <NA> 2020-04-02 alow_rconline 2020-04-02
#> 2   Govt Directed Social Distancing <NA> 2020-04-02 alow_rconline 2020-04-02
#> 3   Govt Directed Social Distancing <NA> 2020-04-02 alow_rconline 2020-04-02
#> 4 Govt Ordered Community Quarantine <NA> 2020-04-02 alow_rconline 2020-04-02
#> 5   Govt Directed Social Distancing <NA> 2020-04-02 alow_rconline 2020-04-02
#> 6 Govt Ordered Community Quarantine <NA> 2020-04-02 alow_rconline 2020-04-02
#>      last_edi_1                             globalid  Shape__Are Shape__Len
#> 1 alow_rconline 60dcdba3-b671-4da9-84fc-9d7c40fda777 48862299113  3299974.8
#> 2 alow_rconline 80d1f2bc-c539-442a-b7f2-71cce0cd700b  4881825512   283774.5
#> 3 alow_rconline a11f6045-d962-4665-8e4a-c6ec14671f67  7683728314   411795.2
#> 4 alow_rconline 13288199-4d9d-493f-9e62-cc1856fde51a 18314912682  2455084.6
#> 5 alow_rconline e642e03e-fe01-4eab-a3f8-a2ab7e432b65 11921048653   596478.7
#> 6 alow_rconline e1dd184d-4804-482e-98ce-11bf1db7bdec 19718437876   604125.1
#>                         geometry
#> 1 MULTIPOLYGON (((-158.7984 5...
#> 2 MULTIPOLYGON (((-96.83707 4...
#> 3 MULTIPOLYGON (((-97.90248 4...
#> 4 MULTIPOLYGON (((-166.6452 5...
#> 5 MULTIPOLYGON (((-100.9647 4...
#> 6 MULTIPOLYGON (((-148.46 61....